src: Fix project mistakes - reward is not calculated correctly

+++ ppo_impl/ppo_torch/wandb/run-20221207_233635-3pvqb5yt/files/code/ppo_impl/ppo_torch/ppo.py
@@ -219,7 +219,7 @@ class PPO_PolicyGradient:
             
             rewards_per_batch.append(rewards_per_episode)
             episode_lengths_per_batch.append(episode + 1) # how long was the episode + 1 as we start at 0
-            rewards_to_go_per_batch = self.rewards_to_go(rewards_per_batch)
+            rewards_to_go_per_batch = self.rewards_to_go(rewards_per_batch, dtype=torch.float)
 
         return torch.tensor(observations_per_batch, dtype=torch.float), \
                 torch.tensor(actions_per_batch, dtype=torch.float), \
